assets_common

Module matching

Structs

FromNetwork

Checks if a is from the expected global consensus network. Checks that Location-a starts with Location-b, and that network is a foreign consensus system.

FromSiblingParachain

Checks if a is from sibling location b. Checks that Location-a starts with Location-b, and that the ParaId of b is not equal to a.

IsForeignConcreteAsset

Accepts an asset if it is from the origin.

LocalLocationPattern

ParentLocation

RemoteAssetFromLocation

Accept an asset if it is native to AssetsAllowedNetworks and it is coming from OriginLocation.