assets_common::runtime_api

Trait FungiblesApi

pub trait FungiblesApi<Block: BlockT, AccountId>: Core<Block>
where
    AccountId: Codec,
{
    // Provided methods
    fn query_account_balances_before_version_2(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        account: AccountId,
    ) -> Result<Result<Vec<Asset>, FungiblesAccessError>, ApiError> { ... }
    fn query_account_balances(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        account: AccountId,
    ) -> Result<Result<VersionedAssets, FungiblesAccessError>, ApiError> { ... }
}

The API for querying account’s balances from runtime.

Provided Methods

fn query_account_balances_before_version_2( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, account: AccountId, ) -> Result<Result<Vec<Asset>, FungiblesAccessError>, ApiError>

👎Deprecated

Returns the list of all Asset that an AccountId has.

fn query_account_balances( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, account: AccountId, ) -> Result<Result<VersionedAssets, FungiblesAccessError>, ApiError>

Returns the list of all Asset that an AccountId has.